rdb: Read a key - read value with expiry

+++ app/timer.c
@@ -3,8 +3,8 @@
 #include "timer.h"
 
 
-long get_time_in_ms() {
+unsigned long get_time_in_ms() {
     struct timeval tv;
     gettimeofday(&tv, NULL);
-    return tv.tv_sec * 1000 + tv.tv_usec / 1000;
+    return (unsigned long)tv.tv_sec * 1000 + (unsigned long)tv.tv_usec / 1000;
 }
